Add Model class
